가사 열기/닫기 기능 구현_예진

+++ src/main/java/com/smhrd/domain/GoodDTO.java
@@ -12,7 +12,6 @@ import lombok.NoArgsConstructor;
 public class GoodDTO {
 
 	private String id;
-	private BigDecimal song_seq;
-	
+	private int song_seq;
 
 }
